Added userid to campaigns and get campaigns by user id
Since we have discord auth on the blazor front end side I don't need to mess around with users or storing information. Instead a campaign will just have the discord id of the user that made it and that'll be used to get campaigns for that user. This'll likely come back to haunt me later in the project. Ideally we'd want to have a table consisting of our user id, external id (discord) and a username we initially pull from discord. Something to think about soon however I'd like to get the mvp out first.